Fix namespace renames.
