+++ BDSMDiscordBot/Models/ExportDateRange.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DiscordImporterBot.Models
+namespace BDSMDiscordBot.Models
 {
     public class ExportDateRange
     {
